lint/format menu files
